misc/HtmlTestSuite/results/test-0924.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>form.elements: namedItem</title>
<link href="mailto:ms2ger@gmail.com" rel="author" title="Ms2ger"/>
<link href="https://html.spec.whatwg.org/multipage/#dom-form-elements" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head><body><div id="log"/>
<div id="test">
<form id="form">
<input id="i1" name="b"/>
<input id="i2" name="b"/>
</form>
</div>
<script>
test(function() {
  assert_true("RadioNodeList" in window, "RadioNodeList should exist");
}, "RadioNodeList should exist")
test(function() {
  var nl = document.forms.form.elements["b"];
  assert_true(nl instanceof NodeList, "Should get a NodeList");
  if ("RadioNodeList" in window) {
    assert_true(nl instanceof RadioNodeList, "Should get a RadioNodeList");
  }
  assert_array_equals(nl,
                      [document.getElementById("i1"),
                       document.getElementById("i2")]);

  var el = nl[0];
  el.parentNode.removeChild(el);
  assert_true(nl instanceof NodeList, "Should get a NodeList");
  if ("RadioNodeList" in window) {
    assert_true(nl instanceof RadioNodeList, "Should get a RadioNodeList");
  }
  assert_array_equals(nl, [document.getElementById("i2")]);
  assert_equals(document.forms.form.elements["b"], document.getElementById("i2"));
}, "elements collection should return elements or RadioNodeLists")
test(function() {
  var fs = document.forms.form.appendChild(document.createElement("fieldset"));
  fs.name = "fs";
  assert_equals(document.forms.form.elements.fs, fs);
  fs.parentNode.removeChild(fs);
}, "elements collection should include fieldsets")
</script>
</body></html>
